+++ src/Clients/OpenAI/OpenAIClient.ts
@@ -1,5 +1,5 @@
 import { ConfigurationOptions, OpenAISettings } from '../../Botc/Configuration/index.js';
-import { ChatCompletionMessageParam } from 'openai/resources/index.mjs';
+import { ChatCompletionMessageParam } from 'openai/resources';
 import { EventBus } from '../../Botc/EventBus/index.js';
 import { Logger } from '../../Botc/Logger/index.js';
 import OpenAI from 'openai';
@@ -7,7 +7,6 @@ import { OpenAINotAllowedError } from './OpenAIClient.errors.js';
 
 /**
  * OpenAI client wrapper
- * @todo Enhance API error handling. Returning '' will break Discord message sending.
  */
 export class OpenAIClient {
   private readonly client: OpenAI;
@@ -22,7 +21,7 @@ export class OpenAIClient {
    */
   constructor(private config: ConfigurationOptions) {
     this.openAIConfig = this.config.llms.openai;
-    this.logger = new Logger(this.config.debugLoggingEnabled.value as boolean);
+    this.logger = new Logger(this.config.featureGates.enableDebugLogging.value as boolean);
 
     this.client = new OpenAI({
       apiKey: this.openAIConfig.apikey.value as string,